requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.4
greenlet==1.0.0
idna==2.10
IMDbPY==2020.9.25
lxml==4.6.3
pytz==2021.1
requests==2.25.1
SQLAlchemy==1.4.5
sqlparse==0.4.1
urllib3==1.26.4
